pyproject.toml
[project]
name = "ynab-butler"
version = "0.1.0"
requires-python = ">=3.12"
description = "YNAB Transaction Importer (Alipay / WeChat)"
readme = "README.md"
dependencies = [
    "pandas>=1.3.0",
    "openpyxl>=3.0.7", # For Excel support
    "python-dotenv>=0.19.0",
    "pytz>=2021.1",
    "requests>=2.26.0",
    "ynab",
    "python-dateutil>=2.8.2",
    "chardet>=4.0.0", # For detecting file encodings
    "pyzipper>=0.3.6",
    "imapclient>=3.0",
    "msal>=1.30.0",
    "pytest>=8.4.2",
    "ruff>=0.14.3",
    "pandas-stubs==2.3.2.250926",
]
